api/requirements.txt
websockets
python-dotenv